+++ finagle-core/src/main/scala/com/twitter/finagle/Path.scala
@@ -16,21 +16,21 @@ import java.util.BitSet
 case class Path(elems: Buf*) {
   require(elems.forall(Path.nonemptyBuf))
 
-  def startsWith(other: Path) = elems.startsWith(other.elems)
+  def startsWith(other: Path): Boolean = elems.startsWith(other.elems)
 
-  def take(n: Int) = Path(elems.take(n): _*)
-  def drop(n: Int) = Path(elems.drop(n): _*)
-  def ++(that: Path) =
+  def take(n: Int): Path = Path(elems.take(n): _*)
+  def drop(n: Int): Path = Path(elems.drop(n): _*)
+  def ++(that: Path): Path =
     if (that.isEmpty) this
     else Path((elems ++ that.elems): _*)
-  def size = elems.size
-  def isEmpty = elems.isEmpty
+  def size: Int = elems.size
+  def isEmpty: Boolean = elems.isEmpty
 
-  lazy val showElems = elems.map(Path.showElem(_))
+  lazy val showElems: Seq[String] = elems.map(Path.showElem(_))
 
-  lazy val show = showElems.mkString("/", "/", "")
+  lazy val show: String = showElems.mkString("/", "/", "")
 
-  override def toString = s"""Path(${showElems.mkString(",")})"""
+  override def toString: String = s"""Path(${showElems.mkString(",")})"""
 }
 
 object Path {